No Bug: Refactored content blocking into better abstracted pieces.
