+++ Client/Frontend/ContentBlocker/ContentBlockerHelper.swift
@@ -10,136 +10,6 @@ import BraveShared
 
 private let log = Logger.browserLogger
 
-// Rename to BlockList
-class BlocklistName: Hashable, CustomStringConvertible {
-    
-    static let ad = BlocklistName(filename: "block-ads")
-    static let tracker = BlocklistName(filename: "block-trackers")
-    static let https = BlocklistName(filename: "upgrade-http")
-    static let image = BlocklistName(filename: "block-images")
-
-    static var allLists: Set<BlocklistName> { return [.ad, .tracker, .https, .image] }
-    
-    let filename: String
-    var rule: WKContentRuleList?
-    
-    init(filename: String) {
-        self.filename = filename
-    }
-    
-    var description: String {
-        return "<\(type(of: self)): \(self.filename)>"
-    }
-    
-    private lazy var fileVersionPref: Preferences.Option<String?>? = {
-        let prefMap = [BlocklistName.ad: Preferences.BlockFileVersion.adblock]
-        return prefMap[self]
-    }()
-    
-    private lazy var fileVersion: String? = {
-        let adVersion = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as? String
-        return self == .ad ? adVersion : nil
-    }()
-    
-    static func blocklists(forDomain domain: Domain) -> (on: Set<BlocklistName>, off: Set<BlocklistName>) {
-        if domain.shield_allOff == 1 {
-            return ([], allLists)
-        }
-        
-        var onList = Set<BlocklistName>()
-        
-        if domain.isShieldExpected(.AdblockAndTp) {
-            onList.formUnion([.ad, .tracker])
-        }
-        
-        // For lists not implemented, always return exclude from `onList` to prevent accidental execution
-        
-        // TODO #159: Setup image shield
-        // TODO #269: Setup HTTPS shield
-        
-        return (onList, allLists.subtracting(onList))
-    }
-
-    private func buildRule(ruleStore: WKContentRuleListStore) -> Deferred<Void> {
-        let compilerDeferred = Deferred<Void>()
-        
-        let compileIt = needsCompiling(ruleStore: ruleStore)
-        compileIt.upon { compile in
-            if !compile {
-                compilerDeferred.fill(())
-                return
-            }
-            
-            BlocklistName.loadJsonFromBundle(forResource: self.filename) { jsonString in
-                ruleStore.compileContentRuleList(forIdentifier: self.filename, encodedContentRuleList: jsonString) { rule, error in
-                    if let error = error {
-                        // TODO #382: Potential telemetry location
-                        log.error("Content blocker '\(self.filename)' errored: \(error.localizedDescription)")
-                        assert(false)
-                    }
-                    assert(rule != nil)
-                    
-                    self.rule = rule
-                    self.fileVersionPref?.value = self.fileVersion
-                    compilerDeferred.fill(())
-                }
-            }
-        }
-        
-        return compilerDeferred
-    }
-    
-    private func needsCompiling(ruleStore: WKContentRuleListStore) -> Deferred<Bool> {
-        let needsCompiling = Deferred<Bool>()
-        if fileVersionPref?.value != fileVersion {
-            // New file, so we must update the lists, no need to check the store
-            needsCompiling.fill(true)
-            return needsCompiling
-        }
-        
-        ruleStore.lookUpContentRuleList(forIdentifier: self.filename) { rule, error in
-            self.rule = rule
-            needsCompiling.fill(self.rule == nil)
-        }
-        return needsCompiling
-    }
-    
-    static func compileAll(ruleStore: WKContentRuleListStore) -> Deferred<Void> {
-        let allCompiledDeferred = Deferred<Void>()
-        let allOfThem = BlocklistName.allLists.map {
-            $0.buildRule(ruleStore: ruleStore)
-        }
-        
-        all(allOfThem).upon { _ in
-            allCompiledDeferred.fill(())
-        }
-        
-        return allCompiledDeferred
-    }
-    
-    private static func loadJsonFromBundle(forResource file: String, completion: @escaping (_ jsonString: String) -> Void) {
-        DispatchQueue.global().async {
-            guard let path = Bundle.main.path(forResource: file, ofType: "json"),
-                let source = try? String(contentsOfFile: path, encoding: .utf8) else {
-                    assert(false)
-                    return
-            }
-            
-            DispatchQueue.main.async {
-                completion(source)
-            }
-        }
-    }
-    
-    public static func == (lhs: BlocklistName, rhs: BlocklistName) -> Bool {
-        return lhs.filename == rhs.filename
-    }
-
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(filename)
-    }
-}
-
 enum BlockerStatus: String {
     case Disabled
     case NoBlockedURLs // When TP is enabled but nothing is being blocked
